Add files via upload
The uploaded code here are versions of a peak analysis program made for data sets with a non-constant baseline. While this could presumably be generalized to other cases, this code was especially designed for analyzing data sets which resulted from researching what are believed to be "ionosomes" using methods of electrochemistry.

I have included here a versions of the code in .ipynb format (one can run such code using Jupyter), as well as in .txt format, so if one chooses to run the code another way, such as using "standard Python", they have the means to easily do that through copy/paste into Python.

I would personally have to recommend using Jupyter if this option is available, as this program was constructed in such a way that it may prove to be easiest to operate using
this software, where individual sections of the code can be executed independently if one wishes (although they still must be run in order from top to bottom typically). As well, the presentation of the code is easily much nicer in Jupyter, and may therefore be much easier for one to follow along with, and understand how the code works and what parameters need to be altered when one uses the code for a new data set.